// backend/src/data/types/data-entity.type.ts
export type SupportedDataEntity =
  | 'users'
  | 'packages'
  | 'invitation-templates'
  | 'package-template-access'
  | 'orders'
  | 'order-items'
  | 'payments'
  | 'user-template-licenses'
  | 'invitations'
  | 'invitation-slugs'
  | 'invitation-contents'
  | 'invitation-media'
  | 'invitation-guests'
  | 'invitation-drafts'
  | 'invitation-feature-grants'
  | 'invitation-visits'
  | 'invitation-rsvps'
  | 'invitation-greetings'
  | 'editor-layout-templates'
  | 'invitation-analytics-daily';